Update Layouts and create Google domain

+++ app/src/main/java/com/study/vipoliveira/tweetanalyst/model/DocumentSentiment.kt
@@ -1,6 +1,6 @@
 package com.study.vipoliveira.tweetanalyst.model
 
 data class DocumentSentiment(
-        val magnitude: Long,
-        val score: Long
+        val magnitude: Double,
+        val score: Double
 )
